+++ src/reduct/api/api_server.h
@@ -26,7 +26,9 @@ class IApiHandler : public IInfoCallback,
                     public IUpdateBucketCallback,
                     public IWriteEntryCallback,
                     public IReadEntryCallback,
-                    public IListEntryCallback {};
+                    public IListEntryCallback,
+                    public IQueryCallback,
+                    public INextCallback {};
 
 /**
  * HTTP API Server