+++ src/reduct/storage/entry.h
@@ -7,18 +7,24 @@
 #include <ostream>
 #include <vector>
 
-#include "reduct/async/io.h"
 #include "reduct/core/error.h"
 #include "reduct/core/result.h"
+#include "reduct/core/time.h"
 #include "reduct/proto/api/entry.pb.h"
+#include "reduct/storage/io/async_io.h"
+#include "reduct/storage/query/quiery.h"
 
 namespace reduct::storage {
 
 /**
  *  Entry of a bucket. Store history of blobs as time series
  */
-class IEntry {
+class IEntry : public io::IAsyncIO, public query::IQuery {
  public:
+  using UPtr = std::unique_ptr<IEntry>;
+  using SPtr = std::shared_ptr<IEntry>;
+  using WPtr = std::weak_ptr<IEntry>;
+
   /**
    * Options
    */
@@ -29,14 +35,12 @@ class IEntry {
     bool operator<=>(const Options& rhs) const = default;
   };
 
-  using Time = std::chrono::system_clock::time_point;
-
   /**
    * Info about a record in a block
    */
   struct RecordInfo {
-    Time time;    // time when it was created
-    size_t size;  // size in bytes
+    core::Time time;  // time when it was created
+    size_t size;      // size in bytes
 
     bool operator<=>(const RecordInfo& rhs) const = default;
 
@@ -44,29 +48,13 @@ class IEntry {
   };
 
   /**
-   * @brief Write a data with timestamp to corresponding block
-   * The method provides the best performance if a new timestamp is always new the stored ones.
-   * Then the engine doesn't need to find a proper block and just records data into the current one.
-   * @param time timestamp of the data
-   * @return async writer or error
-   */
-  [[nodiscard]] virtual core::Result<async::IAsyncWriter::SPtr> BeginWrite(const Time& time, size_t size) = 0;
-
-  /**
-   * @brief Finds the record for the timestamp and read the blob
-   * Current implementation provide only exact matching.
-   * @param time timestamp of record to read
-   * @return async reader or error (404 - if no record found, 500 some internal errors)
-   */
-  [[nodiscard]] virtual core::Result<async::IAsyncReader::SPtr> BeginRead(const Time& time) const = 0;
-
-  /**
    * @brief List records for the time interval [start, stop)
    * @param start
    * @param stop
    * @return return time stamps and size of records,  empty if no data
    */
-  [[nodiscard]] virtual core::Result<std::vector<RecordInfo>> List(const Time& start, const Time& stop) const = 0;
+  [[nodiscard]] virtual core::Result<std::vector<RecordInfo>> List(const core::Time& start,
+                                                                   const core::Time& stop) const = 0;
 
   /**
    * @brief Remove the oldest block from disk
@@ -97,15 +85,7 @@ class IEntry {
    * @param options
    * @return pointer to entre or nullptr if failed to create
    */
-  static std::unique_ptr<IEntry> Build(std::string_view name, const std::filesystem::path& path, Options options);
-
-  /**
-   * @brief Restores entry from dir
-   * Reads files of settings and descriptor
-   * @param full_path path to folder of the entry
-   * @return pointer to entre or nullptr if failed to create
-   */
-  static std::unique_ptr<IEntry> Restore(std::filesystem::path full_path);
+  static IEntry::UPtr Build(std::string_view name, const std::filesystem::path& path, Options options);
 };
 
 }  // namespace reduct::storage